Initial release of the quicksort
A quicksort implementation using three pivot partitioning. It uses an index() function rather than a comparision function and three pivot partitioning because it's tailored to sort pointers to data but it's fast in any case. 

It's pretty blown due to the use of sorting networks for pivot selection and 3 different partitioning schemes one of which is used in daware.