Improved input spec.
daware's input spec was a bit optimized. I'll add a simple stage to construct a SA and ISA from a text that can be fed to daware. Moreover I'll add lyndon array compution. The fact that this simple stage is missing currently is due to the idea of using daware as replacement for trsort in libdivsufsort.